+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE TypeFamilies #-}
+
 module Cardano.Ledger.Babbage.Rules (
   module Cardano.Ledger.Babbage.Rules.Ledger,
   module Cardano.Ledger.Babbage.Rules.Utxo,
@@ -6,6 +9,8 @@ module Cardano.Ledger.Babbage.Rules (
 )
 where
 
+import Cardano.Ledger.Babbage.Core (EraRuleEvent)
+import Cardano.Ledger.Babbage.Era (BabbageEra)
 import Cardano.Ledger.Babbage.Rules.Bbody ()
 import Cardano.Ledger.Babbage.Rules.Deleg ()
 import Cardano.Ledger.Babbage.Rules.Delegs ()
@@ -17,3 +22,10 @@ import Cardano.Ledger.Babbage.Rules.Ppup ()
 import Cardano.Ledger.Babbage.Rules.Utxo
 import Cardano.Ledger.Babbage.Rules.Utxos
 import Cardano.Ledger.Babbage.Rules.Utxow
+import Cardano.Ledger.Shelley.Rules (PpupEvent, ShelleyLedgerEvent, ShelleyTickEvent)
+
+type instance EraRuleEvent "TICK" (BabbageEra c) = ShelleyTickEvent (BabbageEra c)
+
+type instance EraRuleEvent "LEDGER" (BabbageEra c) = ShelleyLedgerEvent (BabbageEra c)
+
+type instance EraRuleEvent "PPUP" (BabbageEra c) = PpupEvent (BabbageEra c)