+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/AdaPots.hs
@@ -43,6 +43,7 @@ import Cardano.Ledger.Shelley.LedgerState.Types (
 import Cardano.Ledger.Shelley.TxBody (unWithdrawals)
 import Cardano.Ledger.UMap (sumRewardsUView)
 import Cardano.Ledger.UTxO (UTxO (..), coinBalance, txInsFilter, txouts)
+import Control.DeepSeq (NFData)
 import Data.Foldable (fold)
 import GHC.Generics (Generic)
 import Lens.Micro ((^.))
@@ -56,6 +57,8 @@ data AdaPots = AdaPots
   , obligationsPot :: Obligations
   }
   deriving (Show, Eq, Generic)
+
+instance NFData AdaPots
 
 -- | Calculate the total ada pots in the epoch state
 totalAdaPotsES ::