+++ eras/babbage/impl/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.7.0.0
 
+* Add type `EraRuleEvent` instances for the event type of:
+  * `UTXOS`
+  * `PPUP`
+  * `LEDGER`
+  * `TICK`
 * Add instances for `InjectRuleFailure` and switch to using `injectFailure`
 * Add `NFData` instance for `BabbageUtxoPredFailure`, `BabbageUtxowPredFailure`
 * Add implementation for `getMinFeeTxUtxo`